Add files via upload
🔥 GOLDEN DRAINER [Update MAY, 2024] | WalletConnect v2.0 | NEW BYPASS | SEAPORT 1.5 | BLUR | X2Y2 | PERMIT | Metamask | Trust Wallet | and more
🚀STABLE OPERATION IS GUARANTEED
Our shop in Telegram: GOLDEN DRAINER - Update MAY, 2024
You can buy automatically this script in Telegram: @verticalio
Golden Drainer is the most advanced drainer on the market. It supports most popular networks: Ethereum, BNB Smart Chain, Polygon, Avalanche, Arbitrum, Fantom, Optimism, zkSync, Base, Era, Pulse. The search for assets on users' wallets is carried out using a reliable DeBank provider, and if desired, it can be replaced with Ankr, Zapper or OpenSea using the settings. Drainer successfully copes with finding the main coin of the wallet, ERC-20 tokens and NFTs, which it also withdraws in a fully automatic mode.

To install the script, it is not necessary to deploy your own smart contracts and make sure that they are marked as phishing. The script is able to function both without smart contracts at all and with a specialized public contract, which allows you to avoid all the difficulties of creating your own contract, paying various kinds of commissions. and connections. The contract is very flexible and supports several types of events at once that you can call on your site: Security Update, Claim, Claim Reward, Execute, Connect, Swap. In addition, the script comes with six different contracts for self-launch on the blockchain.

Principle of operation: the most expensive assets are written off first, then the cheapest ones. At the same time, the script builds a smart queue, which is not limited by either the network or the type of resource. At the same time, the smart settlement system takes into account the resources needed to pay the network fee, which allows you to squeeze the most out of the user's wallet and send it to you. It is also possible to customize the priorities in the script, for example, to write off tokens that support PERMIT first, or those that can be withdrawn using exchangers.

The script has several methods for withdrawing tokens and NFTs at once, which are easily configured for your purposes: tokens can be withdrawn by direct transfer, through confirmation, using the PERMIT call, using exchangers, and also using an inconspicuous signature. Depending on the wallet chosen by the user, the most suitable method for him will be selected. By default, the most inconspicuous and effective is selected. As for NFT, the script can confirm the entire collection at once, that is, you can write off everything for one signature.

Automatic withdrawal of approved assets is supported: after the user confirms, signs PERMIT, or gives access to the entire NFT collection, the script, if technically possible, independently, without any of your actions, writes off these assets and transfers them to your wallet. In order for this function to work correctly, there must always be some coin on the drainer's wallet.

The script supports a huge number of wallets, over 300 to be exact. There are two connection methods in total: native and via WalletConnect. Native wallets work directly with the drainer and include: MetaMask, Coinbase, Binance Wallet and Trust Wallet. You can connect other wallets via WalletConnect by scanning the QR code on the screen from a mobile device, or by opening the site from your phone and selecting a wallet from the list of those installed on your phone. Most mobile wallets support WalletConnect. Works with both versions of WalletConnect.

Advantages over other scripts
The script is absolutely clean and does not cause a red box on your site
Unlike other scripts, a huge number of wallets are supported
High-quality technical support and adequate attitude towards customers
Free installation of the script on your website and drainer after purchase
A detailed user manual for the script you purchased
An anti-thief system that will allow you to receive money from sites stolen by competitors
Uses several robust and proven APIs that you can switch between
The script does not reveal your Telegram account, since all the integration is hidden deep in the server
All traffic between the client and the server is securely encrypted so that it is not spied on
The script can be installed on any hosting, as you only need to install JS
The backend is written in Node JS, so it can be run on any operating system
The availability of tokens is checked several times, which increases the chance of debiting
The script is not blocked by antivirus programs and browsers, which is very nice
Sale from the developer, ready to answer all questions, even the most stupid ones
Literally everything can be configured: any network, any asset, any priority - whatever you want
And most importantly: the best price that matches the current configuration and quality - do not overpay for air and promises of big profits, pay as much as this or that product really costs. By purchasing a script from me, you are investing in the future, namely: you get a high-quality code, free fixes, the fastest technical support that will chew you literally everything
Description of the main functionality of the script
Supported networks: Ethereum, BNB SmartChain, Polygon, Avalanche, Arbitrum, Fantom, Optimism, zkSync, Base, Era, Pulse

Supported Assets: Native Coins, Tokens and NFT

Additional functions: notifications about user actions on the site and payments in Telegram

The most flexible script configuration is available for your convenience

Native coin withdrawal method: Sign / Transfer / Smart Contract

Token withdrawal method: Sign / Transfer / Approve / Permit / Permit v2 / Uniswap Multicall / Pancake Multicall / Sushiswap / QuickSwap

NFT withdrawal method: Sign / Transfer / SetApprovalForAll / SeaPort 1.5 / Blur Private Exploit / X2Y2 Private Exploit

To work, you will need a clean virtual server on Ubuntu 22.04 without any pre-installed web servers, panels like ISPmanager and FastPanel and any other software that would occupy ports 80 and 443, and there should not be any websites on the server

Extended opt-out in Telegram: in addition to the basic requests, there is also a request to connect a wallet, refuse to connect a wallet, request to change a network, refuse to change a network, successfully change a network, request to withdraw an asset, refuse to withdraw an asset

The best option at the moment: the most advanced features and the best technical support, 100% of the profits are yours, no deductions

Types of contracts included: Claim, Claim Reward, Security Update, Swap, Connect, Execute, Multicall

Available trust contract format: 0x000...0000 - maximum user trust

Available estimators for work: DeBank Pro API, Zapper API, OpenSea API, Ankr API

You can buy this script in our Telegram shop automatically: @Drainersdepot_bot

Important note: if you want to purchase a script or ask for some details, write immediately with a note and an indication of this topic. Messages without an indication of the purpose of the appeal have a 99.9% chance of being ignored. Let's save each other's time.

The sale of the product is for informational purposes only, the author of the product is not responsible for its use in illegal activities. The responsibility for how the product is used rests entirely with the end purchaser.

 

If you have any questions: @verticalio